backend/geofinder/requirements.txt
spacy
requests
python-dotenv
openai
tqdm
python-dotenv
google-cloud-firestore